% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{Earthquakes data}
\format{
An object of class \code{ts} of length 107.
}
\source{
\url{http://www.hmms-for-time-series.de/second/data/earthquakes.txt}
}
\usage{
earthquakes
}
\description{
Earthquakes data
}
\keyword{datasets}
