% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshmm_class.R
\name{tsHMM}
\alias{tsHMM}
\title{Forecasting Using Hidden Markov Models}
\source{
\url{http://www.hmms-for-time-series.de/second/appendix-a/ZMLcode.txt}
}
\description{
Forecasting Using Hidden Markov Models
}
\examples{
\dontrun{

library(tsHMM)

data("earthquakes", package = "tsHMM")

m = 2

lambda <- c(15,25)

gamma <- matrix(c(0.9,0.1,0.1,0.9), m, m,byrow=TRUE)

model <- tsHMM$new(x = earthquakes, h = 12, m = 2, lambda = lambda, gamma = gamma)

model$fit()

fc <- model$forecast(xf = 0:50)

fc

}
}
\references{
Zucchini, W., MacDonald, I.L., & Langrock, R. (2016).
Hidden Markov Models for Time Series: An Introduction Using R (2nd ed.).
Chapman and Hall/CRC. https://doi.org/10.1201/b20790
}
\author{
Resul Akay
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tsHMM$new()}}
\item \href{#method-fit}{\code{tsHMM$fit()}}
\item \href{#method-forecast}{\code{tsHMM$forecast()}}
\item \href{#method-plot}{\code{tsHMM$plot()}}
\item \href{#method-clone}{\code{tsHMM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a HMM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tsHMM$new(x, h = frequency(x), m, lambda, gamma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A univariate time series}

\item{\code{h}}{Forecast horizon}

\item{\code{m}}{Number of stats}

\item{\code{lambda}}{lambda}

\item{\code{gamma}}{gamma}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a HMM to a univariate timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tsHMM$fit(delta = NULL, stationary = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{delta}

\item{\code{stationary}}{Bool. If stationary = TRUE data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forecast"></a>}}
\if{latex}{\out{\hypertarget{method-forecast}{}}}
\subsection{Method \code{forecast()}}{
Forecast fitted HMM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tsHMM$forecast(xf = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xf}}{xf}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tsHMM$plot(h = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tsHMM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
